Use an integer as offset

+++ tests/Doctrine/ODM/MongoDB/Tests/Functional/MemoryUsageTest.php
@@ -61,6 +61,6 @@ class MemoryUsageTest extends BaseTest
     {
         $unit = ['b', 'kb', 'mb', 'gb', 'tb', 'pb'];
 
-        return round($size / pow(1024, ($i = floor(log($size, 1024)))), 2) . ' ' . $unit[$i];
+        return round($size / pow(1024, ($i = (int) floor(log($size, 1024)))), 2) . ' ' . $unit[$i];
     }
 }
